apq8064: Fix baseband version
The baseband version string is set by both RIL and this script.  But RIL is
setting the empty string for some reason.  Both of these processes run
independently and race with each other.  Fix the race by waiting up to 10
seconds after boot complete for RIL to set the baseband version string,
then overwrite it with the correct string.

Change-Id: Ic9046c0a2c6b2bba1666fa1acb6fb370173fb59a